pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sql-analytics-workspace"
version = "0.1.0"
description = "OpenEnv environment for SQL analytics — agent writes queries to answer business questions"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["server*", "models*", "client*"]
